+++ CHANGELOG.md
@@ -7,7 +7,10 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 ### Added
 
 * `FixedDecimal` Haskell equivalent to `PFixedDecimal`, along with `Num` and
-* `Fractional` instances, `fixedNumerator` and `fixedDenominator` extractor functions.
+* `Fractional` instances and the following functions:
+  - `fixedNumerator`
+  - `fixedDenominator`
+  - `convertExp`
 
 ### Modified