+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 * `Fractional` instances and the following functions:
   - `fixedNumerator`
   - `fixedDenominator`
+  - `emul`
+  - `ediv`
   - `convertExp`
 
 ### Modified
@@ -20,6 +22,8 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
   - `PFixedDecimal`
   - `pfixedNumerator`
   - `pfixedDenominator`
+  - `pemul`
+  - `pediv`
   - `pconvertExp`
   - `pfromFixedDecimal`
   - `ptoFixedDecimal`